# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AutoManager POS API
  version: "1.0.0"
  description: OpenAPI definition for AutoManager POS (core endpoints)
servers:
  - url: https://api.example.com
    description: Primary server (replace with real URL)
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Idempotency-Key:
      name: Idempotency-Key
      in: header
      description: Client-generated idempotency key for create operations
      required: false
      schema:
        type: string
  schemas:
    Error:
      type: object
      properties:
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        errors:
          type: array
          items:
            type: object
...
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        sku:
          type: string
        price:
          type: number
          format: float
    OrderItem:
      type: object
      required: [product_id,quantity,unit_price]
      properties:
        id:
          type: string
          format: uuid
        product_id:
          type: string
          format: uuid
        name:
          type: string
        quantity:
          type: integer
        unit_price:
          type: number
          format: float
        subtotal:
          type: number
    Payment:
      type: object
      required: [amount,payment_method]
      properties:
        id:
          type: string
          format: uuid
        amount:
          type: number
          format: float
        payment_method:
          type: string
          enum: [CASH, CARD, QR, WALLET, OTHER]
        provider_metadata:
          type: object
    OrderCreate:
      type: object
      required: [branch_id,order_type,items]
      properties:
        client_id:
          type: string
          description: local temporary id from client
        branch_id:
          type: string
          format: uuid
        created_by:
          type: string
          format: uuid
        order_type:
          type: string
          enum: [DINE_IN, TAKEAWAY]
        table_id:
          type: string
          format: uuid
          nullable: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        payments:
          type: array
          items:
            $ref: '#/components/schemas/Payment'
        metadata:
          type: object
    OrderResponse:
      allOf:
        - $ref: '#/components/schemas/OrderCreate'
        - type: object
          properties:
            id:
              type: string
              format: uuid
            order_code:
              type: string
            order_status:
              type: string
            payment_status:
              type: string
            total_amount:
              type: number
              format: float
            created_at:
              type: string
              format: date-time
...
paths:
...
  /orders:
    post:
      summary: Create an order
      parameters:
        - $ref: '#/components/parameters/Idempotency-Key'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreate'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '200':
          description: Duplicate idempotent submission — existing order returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: List orders (basic)
      parameters:
        - name: branch_id
          in: query
          schema:
            type: string
            format: uuid
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderResponse'
...
  /orders/{order_id}:
    parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get order by id
      responses:
        '200':
          description: Order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '404':
          description: Not found
    patch:
      summary: Update order (partial)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Updated order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '409':
          description: Conflict
  /orders/{order_id}/payments:
    post:
      summary: Create payment for an order
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Payment'
      responses:
        '201':
          description: Payment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '400':
          description: Validation error
  /orders/{order_id}/items:
    post:
      summary: Add order item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderItem'
      responses:
        '201':
          description: Item added
  /orders/{order_id}/items/{item_id}:
    parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: item_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      summary: Update order item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                quantity:
                  type: integer
                unit_price:
                  type: number
      responses:
        '200':
          description: Updated
    delete:
      summary: Delete order item
      responses:
        '200':
          description: Deleted
...
  /products:
    get:
      summary: List products (optionally by branch)
      parameters:
        - name: branch_id
          in: query
          schema:
            type: string
            format: uuid
        - name: category_id
          in: query
          schema:
            type: string
            format: uuid
        - name: q
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Product list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
    post:
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [branch_id, name, price]
              properties:
                branch_id:
                  type: string
                  format: uuid
                category_id:
                  type: string
                  format: uuid
                sku:
                  type: string
                name:
                  type: string
                price:
                  type: number
                metadata:
                  type: object
      responses:
        '201':
          description: Product created
  /products/{product_id}:
    parameters:
      - name: product_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      summary: Update product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Product updated
    delete:
      summary: Delete product
      responses:
        '200':
          description: Product deleted
  /product-categories:
    get:
      summary: List product categories
      responses:
        '200':
          description: Category list
    post:
      summary: Create product category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        '201':
          description: Category created
...
  /shifts:
      get:
        summary: List shifts
        responses:
          '200':
            description: Shifts list
      post:
        summary: Create shift
        requestBody:
          required: true
          content:
            application/json:
              schema:
                type: object
                required: [name, start_time, end_time]
                properties:
                  name:
                    type: string
                  start_time:
                    type: string
                    example: "09:00"
                  end_time:
                    type: string
                    example: "16:00"
        responses:
          '201':
            description: Shift created
...
  /attendance/checkin:
      post:
        summary: Attendance check-in
        requestBody:
          required: true
          content:
            application/json:
              schema:
                type: object
                required: [employee_id, shift_id]
                properties:
                  employee_id:
                    type: string
                    format: uuid
                  shift_id:
                    type: string
                    format: uuid
        responses:
          '201':
            description: Checked in
  /attendance/checkout:
      post:
        summary: Attendance check-out
        requestBody:
          required: true
          content:
            application/json:
              schema:
                type: object
                required: [employee_id]
                properties:
                  employee_id:
                    type: string
                    format: uuid
        responses:
          '200':
            description: Checked out
  /tables:
      get:
        summary: List tables
        parameters:
          - name: branch_id
            in: query
            schema:
              type: string
              format: uuid
        responses:
          '200':
            description: Tables list
      post:
        summary: Create table
        requestBody:
          required: true
          content:
            application/json:
              schema:
                type: object
                required: [branch_id, name]
                properties:
                  branch_id:
                    type: string
                    format: uuid
                  name:
                    type: string
                  status:
                    type: string
                    enum: [AVAILABLE, OCCUPIED]
        responses:
          '201':
            description: Table created
  /tables/{table_id}:
      parameters:
        - name: table_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      patch:
        summary: Update table
        requestBody:
          required: true
          content:
            application/json:
              schema:
                type: object
        responses:
          '200':
            description: Table updated
      delete:
        summary: Delete table
        responses:
          '200':
            description: Table deleted
  /tables/{table_id}/status:
      patch:
        summary: Update table status
        parameters:
          - name: table_id
            in: path
            required: true
            schema:
              type: string
              format: uuid
        requestBody:
          required: true
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    enum: [AVAILABLE, OCCUPIED]
        responses:
          '200':
            description: Updated
...